# openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "description" : "Clinic management API. Every request runs inside the clinic bound to the caller's session,\nand reads never cross that boundary.\n\nAuthentication is a session cookie: `POST /api/session` opens it, `DELETE /api/session` closes it,\nand every other operation requires it.\n\nEndpoints belonging to an optional module answer `404` while the clinic has not activated that\nmodule, so a client should treat their absence as a feature the clinic did not contract.\n`GET /api/capabilities` reports what is currently available.\n\nErrors always carry the same body: status, message, request path and, for a rejected payload,\nthe offending fields.",
    "title" : "Clivo API",
    "version" : "0.0.1"
  },
  "servers" : [ {
    "url" : "http://localhost",
    "description" : "Generated server url"
  } ],
  "security" : [ {
    "sessionCookie" : [ ]
  } ],
...
  "paths" : {
...
    "/api/commission-rates" : {
      "get" : {
        "operationId" : "listCommissionRates",
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/RateView"
                  }
                }
              }
            },
            "description" : "OK"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Unauthorized"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Forbidden"
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Not Found"
          },
          "422" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Unprocessable Content"
          }
        },
        "summary" : "List the practitioners' rates",
        "tags" : [ "Commissions" ]
      }
    },
    "/api/commissions" : {
      "get" : {
        "operationId" : "getCommissionStatement",
        "parameters" : [ {
          "in" : "query",
          "name" : "year",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        }, {
          "in" : "query",
          "name" : "month",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/StatementView"
                }
              }
            },
            "description" : "OK"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Unauthorized"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Forbidden"
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Not Found"
          },
          "422" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Unprocessable Content"
          }
        },
        "summary" : "Report each practitioner's share for a month",
        "tags" : [ "Commissions" ]
      }
    },
    "/api/commissions/closing" : {
      "post" : {
        "operationId" : "closeCommissionPeriod",
        "parameters" : [ {
          "in" : "query",
          "name" : "year",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        }, {
          "in" : "query",
          "name" : "month",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/StatementView"
                }
              }
            },
            "description" : "OK"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Unauthorized"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Forbidden"
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Not Found"
          },
          "422" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Unprocessable Content"
          }
        },
        "summary" : "Close a month, freezing what is owed",
        "tags" : [ "Commissions" ]
      }
    },
...
    "/api/practitioners/{practitionerId}/commission-rate" : {
      "put" : {
        "operationId" : "setCommissionRate",
        "parameters" : [ {
          "in" : "path",
          "name" : "practitionerId",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/RateRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RateView"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Bad Request"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Unauthorized"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Forbidden"
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Not Found"
          },
          "422" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Unprocessable Content"
          }
        },
        "summary" : "Set the rate a practitioner is paid at",
        "tags" : [ "Commissions" ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "CommissionView" : {
        "type" : "object",
        "properties" : {
          "amount" : {
            "type" : "number"
          },
          "encounterId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "percentage" : {
            "type" : "number"
          },
          "practitionerId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "practitionerName" : {
            "type" : "string"
          },
          "status" : {
            "type" : "string"
          }
        }
      },
...
      "ErrorResponse" : {
        "type" : "object",
        "properties" : {
          "error" : {
            "type" : "string"
          },
          "message" : {
            "type" : "string"
          },
          "path" : {
            "type" : "string"
          },
          "status" : {
            "type" : "integer",
            "format" : "int32"
          },
          "timestamp" : {
            "type" : "string",
            "format" : "date-time"
          },
          "violations" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/FieldViolation"
            }
          }
        }
      },
...
      "FieldViolation" : {
        "type" : "object",
        "properties" : {
          "field" : {
            "type" : "string"
          },
          "message" : {
            "type" : "string"
          }
        }
      },
...
      "RateRequest" : {
        "type" : "object",
        "properties" : {
          "percentage" : {
            "type" : "number"
          }
        },
        "required" : [ "percentage" ]
      },
      "RateView" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "percentage" : {
            "type" : "number"
          },
          "practitionerId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "practitionerName" : {
            "type" : "string"
          }
        }
      },
...
      "StatementView" : {
        "type" : "object",
        "properties" : {
          "closed" : {
            "type" : "boolean"
          },
          "commissions" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/CommissionView"
            }
          },
          "period" : {
            "type" : "string"
          },
          "total" : {
            "type" : "number"
          }
        }
      },
...
    },
    "securitySchemes" : {
      "sessionCookie" : {
        "description" : "Session cookie issued by POST /api/session.",
        "in" : "cookie",
        "name" : "JSESSIONID",
        "type" : "apiKey"
      }
    }
  }
}